+++ packages/orq-rc/RELEASES.md
@@ -4919,3 +4919,13 @@ Based on:
 - [typescript v4.0.0-rc.23] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.23] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.23 - packages/orq-rc
+
+## 2025-11-12 05:39:39
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.654.0 (2.750.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.24] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.24] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.24 - packages/orq-rc